Moved VLC library to GUI
